# parser: optimize performance