--- a/yy_parser.go
+++ b/yy_parser.go
@@ -100,7 +100,7 @@ type Parser struct {
 	// the following fields are used by yyParse to reduce allocation.
 	cache  []yySymType
 	yylval yySymType
-	yyVAL  yySymType
+	yyVAL  *yySymType
 }
 
 type stmtTexter interface {