### What problem does this PR solve?
Optimizing parser performance

### What is changed and how it works?
Here is the benchmark, 8%~12% faster than before:

goos: linux
goarch: amd64

Before optimization:
```sh
BenchmarkSysbenchSelect       200000          6121 ns/op        1824 B/op         20 allocs/op
BenchmarkParseComplex          10000        223586 ns/op        61772 B/op        642 allocs/op
BenchmarkParseSimple          100000         19047 ns/op        5144 B/op         71 allocs/op
```
After optimization:
```sh
BenchmarkSysbenchSelect       200000          5586 ns/op        1824 B/op         20 allocs/op
BenchmarkParseComplex          10000        206988 ns/op        61772 B/op        642 allocs/op
BenchmarkParseSimple          100000         16873 ns/op        5144 B/op         71 allocs/op
```

### Check List
Tests

Unit test